+++ ShipTrip/Views/Cruises/CruiseTimelineRowView.swift
@@ -35,7 +35,10 @@ struct CruiseTimelineRowView: View {
                 }
             }
         }
-        .padding(.vertical, 6)
+        .padding(.horizontal, 12)
+        .padding(.vertical, 10)
+        .background(Color(UIColor.secondarySystemBackground))
+        .clipShape(RoundedRectangle(cornerRadius: 10))
     }
 
     // MARK: Subviews